+++ patches/tModLoader/Terraria/ModLoader/ModItem.cs
@@ -313,7 +313,7 @@ namespace Terraria.ModLoader
 		/// </summary>
 		/// <param name="player">The player using the item.</param>
 		/// <param name="crit">The total crit chance of the item after all normal crit chance calculations.</param>
-		public virtual void ModifyWeaponCrit(Player player, ref int crit) {
+		public virtual void ModifyWeaponCrit(Player player, ref float crit) {
 		}
 
 		/// <summary>
